+++ src/test/db/badge-index.test.ts
@@ -5,8 +5,7 @@ import { badgeRequirementDataFixture } from '../api/badge-requirement-data.fixtu
 describe(BadgeIndex.name, () => {
   describe('Constructor', () => {
     test(`should throw an error on duplicate key`, () => {
-      const index = new BadgeIndex()
-      expect(() => index.load([
+      expect(() => new BadgeIndex([
         new Badge(badgeDataFixture.create({ key: 'foo' })),
         new Badge(badgeDataFixture.create({ key: 'foo' })),
       ])).toThrow('Duplicate key [foo]')
@@ -15,27 +14,23 @@ describe(BadgeIndex.name, () => {
 
   describe('get', () => {
     test(`should retrieve badge from the index`, () => {
-      const data = [new Badge(badgeDataFixture.create({ key: 'foo' }))]
-
-      const index = new BadgeIndex()
-      index.load(data)
+      const index = new BadgeIndex([new Badge(badgeDataFixture.create({ key: 'foo' }))])
       expect(index.get('foo')).not.toBeUndefined()
     })
 
     test(`should return undefined for unknown badge`, () => {
-      expect(new BadgeIndex().get('foo')).toBeUndefined()
+      expect(new BadgeIndex([]).get('foo')).toBeUndefined()
     })
 
     test(`should return undefined for undefined key`, () => {
       const key = undefined
-      expect(new BadgeIndex().get(key)).toBeUndefined()
+      expect(new BadgeIndex([]).get(key)).toBeUndefined()
     })
   })
 
   describe('search', () => {
     test(`should return everything for an empty query`, () => {
-      const index = new BadgeIndex()
-      index.load([
+      const index = new BadgeIndex([
         new Badge(badgeDataFixture.create({ key: 'foo-1', acquisition: 'Foo 1' })),
         new Badge(badgeDataFixture.create({ key: 'foo-2', acquisition: 'Foo 2' })),
         new Badge(badgeDataFixture.create({ key: 'bar-1', acquisition: 'Bar 1' })),
@@ -51,8 +46,7 @@ describe(BadgeIndex.name, () => {
 
     describe('query', () => {
       test(`should match on badge name`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'match-1', name: [{ value: 'Foo 1' }] })),
           new Badge(badgeDataFixture.create({ key: 'match-2', name: [{ value: 'Foo 2' }, { value: 'Bar 2' }] })),
           new Badge(badgeDataFixture.create({ key: 'match-3', name: [{ value: 'Bar 3' }, { value: 'Foo 3' }] })),
@@ -65,8 +59,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should match on badge text`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'match-1', badgeText: [{ value: 'Foo 1' }] })),
           new Badge(badgeDataFixture.create({ key: 'match-2', badgeText: [{ value: 'Foo 2' }, { value: 'Bar 2' }] })),
           new Badge(badgeDataFixture.create({ key: 'match-3', badgeText: [{ value: 'Bar 3' }, { value: 'Foo 3' }] })),
@@ -80,8 +73,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should match on acquisition`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'match-1', acquisition: 'Foo 1' })),
           new Badge(badgeDataFixture.create({ key: 'match-2', acquisition: 'Foo 2' })),
           new Badge(badgeDataFixture.create({ key: 'miss-1', acquisition: 'Bar 1' })),
@@ -94,8 +86,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should match on effect`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'match-1', effect: 'Foo 1' })),
           new Badge(badgeDataFixture.create({ key: 'match-2', effect: 'Foo 2' })),
           new Badge(badgeDataFixture.create({ key: 'miss-1', effect: 'Bar 1' })),
@@ -108,8 +99,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should match on notes`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'match-1', notes: 'Foo 1' })),
           new Badge(badgeDataFixture.create({ key: 'match-2', notes: 'Foo 2' })),
           new Badge(badgeDataFixture.create({ key: 'miss-1', notes: 'Bar 1' })),
@@ -122,8 +112,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should match on setTitle`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'match-1', setTitleId: [123] })),
           new Badge(badgeDataFixture.create({ key: 'match-2', setTitleId: [456, 123] })),
           new Badge(badgeDataFixture.create({ key: 'miss-1', setTitleId: [456] })),
@@ -138,8 +127,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should match the start of a string`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'match-1', acquisition: 'Foo 1' })),
           new Badge(badgeDataFixture.create({ key: 'match-2', acquisition: 'Foo 2' })),
           new Badge(badgeDataFixture.create({ key: 'miss-1', acquisition: 'Bar 1' })),
@@ -151,8 +139,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should be case insensitive`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'match-1', acquisition: 'Foo 1' })),
           new Badge(badgeDataFixture.create({ key: 'match-2', acquisition: 'Foo 2' })),
           new Badge(badgeDataFixture.create({ key: 'miss-1', acquisition: 'Bar 1' })),
@@ -164,8 +151,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should default to querying on name only`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'match-1', name: [{ value: 'Foo 1' }] })),
           new Badge(badgeDataFixture.create({ key: 'miss-1', acquisition: 'Foo 2' })),
           new Badge(badgeDataFixture.create({ key: 'miss-2', name: [{ value: 'Bar 1' }] })),
@@ -180,8 +166,7 @@ describe(BadgeIndex.name, () => {
 
     describe('filter', () => {
       test(`should filter nothing if not specified`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1' })),
           new Badge(badgeDataFixture.create({ key: 'badge-2' })),
           new Badge(badgeDataFixture.create({ key: 'badge-3' })),
@@ -196,8 +181,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should filter on badge type`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1', type: 'exploration' })),
           new Badge(badgeDataFixture.create({ key: 'badge-2', type: 'exploration' })),
           new Badge(badgeDataFixture.create({ key: 'badge-3', type: 'history' })),
@@ -212,8 +196,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should filter on badge zone`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1', requirements: [{ location: { zoneKey: 'atlas-park' } }] })),
           new Badge(badgeDataFixture.create({ key: 'badge-2', requirements: [{ location: { zoneKey: 'perez-park' } }] })),
           new Badge(badgeDataFixture.create({ key: 'badge-3', requirements: [{ location: { zoneKey: 'abandoned-sewer-network' } }] })),
@@ -233,8 +216,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should filter on alignment`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1', morality: ['hero'] })),
           new Badge(badgeDataFixture.create({ key: 'badge-2', morality: ['villain'] })),
           new Badge(badgeDataFixture.create({ key: 'badge-3', morality: ['loyalist'] })),
@@ -251,8 +233,7 @@ describe(BadgeIndex.name, () => {
 
     describe('pagination', () => {
       test(`should return all results with no pagination data`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1' })),
           new Badge(badgeDataFixture.create({ key: 'badge-2' })),
           new Badge(badgeDataFixture.create({ key: 'badge-3' })),
@@ -267,13 +248,12 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should be 1-based for page number`, () => {
-        const result = new BadgeIndex().search()
+        const result = new BadgeIndex([]).search()
         expect(result.page).toBe(1)
       })
 
       test(`should return the requested page size`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1' })),
           new Badge(badgeDataFixture.create({ key: 'badge-2' })),
           new Badge(badgeDataFixture.create({ key: 'badge-3' })),
@@ -287,8 +267,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should return the start of the array with no page specified`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1' })),
           new Badge(badgeDataFixture.create({ key: 'badge-2' })),
           new Badge(badgeDataFixture.create({ key: 'badge-3' })),
@@ -303,8 +282,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should return results from the middle of the array with a page specified`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1' })),
           new Badge(badgeDataFixture.create({ key: 'badge-2' })),
           new Badge(badgeDataFixture.create({ key: 'badge-3' })),
@@ -319,8 +297,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should return a partial page if at the end of the array`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1' })),
           new Badge(badgeDataFixture.create({ key: 'badge-2' })),
           new Badge(badgeDataFixture.create({ key: 'badge-3' })),
@@ -334,8 +311,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should return the correct total entry count`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1' })),
           new Badge(badgeDataFixture.create({ key: 'badge-2' })),
           new Badge(badgeDataFixture.create({ key: 'badge-3' })),
@@ -348,8 +324,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should return the page size`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1' })),
           new Badge(badgeDataFixture.create({ key: 'badge-2' })),
           new Badge(badgeDataFixture.create({ key: 'badge-3' })),
@@ -362,8 +337,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should return the correct total page count`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1' })),
           new Badge(badgeDataFixture.create({ key: 'badge-2' })),
           new Badge(badgeDataFixture.create({ key: 'badge-3' })),
@@ -376,8 +350,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should return a total page count of 1 when no page size is provided`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1' })),
           new Badge(badgeDataFixture.create({ key: 'badge-2' })),
           new Badge(badgeDataFixture.create({ key: 'badge-3' })),
@@ -390,8 +363,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should return the last page if a page is requested past the max`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1' })),
           new Badge(badgeDataFixture.create({ key: 'badge-2' })),
           new Badge(badgeDataFixture.create({ key: 'badge-3' })),
@@ -406,15 +378,14 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should return the first page if a page is requested lower than 1`, () => {
-        const result = new BadgeIndex().search({ page: -10 })
+        const result = new BadgeIndex([]).search({ page: -10 })
         expect(result.page).toBe(1)
       })
     })
 
     describe('sort', () => {
       test(`should not modify sort if not specified`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1' })),
           new Badge(badgeDataFixture.create({ key: 'badge-2' })),
           new Badge(badgeDataFixture.create({ key: 'badge-3' })),
@@ -427,8 +398,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should not modify sort if order is canonical`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1' })),
           new Badge(badgeDataFixture.create({ key: 'badge-2' })),
           new Badge(badgeDataFixture.create({ key: 'badge-3' })),
@@ -441,8 +411,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should reverse default sort with desc`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1' })),
           new Badge(badgeDataFixture.create({ key: 'badge-2' })),
           new Badge(badgeDataFixture.create({ key: 'badge-3' })),
@@ -455,8 +424,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should sort by badge name`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1', name: [{ value: 'Abc' }] })),
           new Badge(badgeDataFixture.create({ key: 'badge-2', name: [{ value: 'XYZ' }] })),
           new Badge(badgeDataFixture.create({ key: 'badge-3', name: [{ value: 'AAB' }] })),
@@ -468,8 +436,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should sort by badge name descending`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1', name: [{ value: 'Abc' }] })),
           new Badge(badgeDataFixture.create({ key: 'badge-2', name: [{ value: 'XYZ' }] })),
           new Badge(badgeDataFixture.create({ key: 'badge-3', name: [{ value: 'AAB' }] })),
@@ -481,8 +448,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should use the default badge name when sorting by name`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1', name: [{ value: 'Abc' }] })),
           new Badge(badgeDataFixture.create({ key: 'badge-2', name: [{ value: 'XYZ' }, { sex: 'F', value: 'AAA' }] })),
           new Badge(badgeDataFixture.create({ key: 'badge-3', name: [{ value: 'AAB' }] })),
@@ -494,8 +460,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should sort by zone name`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1', requirements: [{ location: { zoneKey: 'atlas-park' } }] })),
           new Badge(badgeDataFixture.create({ key: 'badge-2', requirements: [{ location: { zoneKey: 'perez-park' } }] })),
           new Badge(badgeDataFixture.create({ key: 'badge-3', requirements: [{ location: { zoneKey: 'abandoned-sewer-network' } }] })),
@@ -507,8 +472,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should sort by zone name descending`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1', requirements: [{ location: { zoneKey: 'atlas-park' } }] })),
           new Badge(badgeDataFixture.create({ key: 'badge-2', requirements: [{ location: { zoneKey: 'perez-park' } }] })),
           new Badge(badgeDataFixture.create({ key: 'badge-3', requirements: [{ location: { zoneKey: 'abandoned-sewer-network' } }] })),
@@ -520,8 +484,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should maintain canonical as secondary sort when sorting by zone name`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1', requirements: [{ location: { zoneKey: 'atlas-park' } }] })),
           new Badge(badgeDataFixture.create({ key: 'badge-2', requirements: [{ location: { zoneKey: 'perez-park' } }] })),
           new Badge(badgeDataFixture.create({ key: 'badge-3', requirements: [{ location: { zoneKey: 'atlas-park' } }] })),
@@ -534,8 +497,7 @@ describe(BadgeIndex.name, () => {
       })
 
       test(`should sort undefined or multiple zone names to the end`, () => {
-        const index = new BadgeIndex()
-        index.load([
+        const index = new BadgeIndex([
           new Badge(badgeDataFixture.create({ key: 'badge-1', requirements: [{ location: { zoneKey: 'atlas-park' } }] })),
           new Badge(badgeDataFixture.create({ key: 'badge-2', requirements: [{ location: undefined }] })),
           new Badge(badgeDataFixture.create({ key: 'badge-3', requirements: [{ location: { zoneKey: 'perez-park' } }] })),