+++ src/test/db/coh-content-database.test.ts
@@ -7,53 +7,27 @@ import { missionDataFixture } from '../api/mission-data.fixture'
 import { bundleDataFixture } from '../api/bundle-data.fixture'
 
 describe(CohContentDatabase.name, () => {
-  describe('load', () => {
+  describe('Constructor', () => {
     test('should load a basic bundle', () => {
-      const database = new CohContentDatabase()
-      database.load({})
-    })
-
-    test('should reset the database if load is called again', () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture.create({
-        header: { name: 'Homecoming' },
-        servers: ['Test'],
-        archetypes: [archetypeDataFixture.create()],
-        zones: [zoneDataFixture.create()],
-        contacts: [contactDataFixture.create()],
-        missions: [missionDataFixture.create()],
-        badges: [badgeDataFixture.create()],
-      }))
-
-      expect(database.header?.name).toEqual('Homecoming')
-      expect(database.servers).toHaveLength(1)
-      expect(database.archetypes).toHaveLength(1)
-      expect(database.zones).toHaveLength(1)
-      expect(database.contacts).toHaveLength(1)
-      expect(database.missions).toHaveLength(1)
-      expect(database.badges).toHaveLength(1)
-
-      database.load(bundleDataFixture.create({ header: { name: 'Reset' } }))
-      expect(database.header?.name).toEqual('Reset')
-      expect(database.servers).toHaveLength(0)
-      expect(database.archetypes).toHaveLength(0)
-      expect(database.zones).toHaveLength(0)
-      expect(database.contacts).toHaveLength(0)
-      expect(database.missions).toHaveLength(0)
-      expect(database.badges).toHaveLength(0)
+      new CohContentDatabase(bundleDataFixture.create())
     })
   })
 
   describe('header', () => {
-    test(`should be undefined if not initialized`, () => {
-      const database = new CohContentDatabase()
-      expect(database.header).toBeUndefined()
+    test(`should accept an undefined field`, () => {
+      const database = new CohContentDatabase(
+        bundleDataFixture.omit('header').create(),
+      )
+      expect(database.header).toBeDefined()
+      expect(database.header?.name).toBeUndefined()
     })
 
     test(`should load values from bundle`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture
-        .create({ header: { name: 'Homecoming' } }))
+      const database = new CohContentDatabase(
+        bundleDataFixture.create(
+          { header: { name: 'Homecoming' } },
+        ),
+      )
 
       expect(database.header?.name).toBe('Homecoming')
     })
@@ -61,153 +35,127 @@ describe(CohContentDatabase.name, () => {
 
   describe('servers', () => {
     test(`should accept an undefined field`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture
-        .omit('servers')
-        .create())
-
-      expect(database.servers).toHaveLength(0)
-    })
-
-    test(`should be empty if uninitialized`, () => {
-      const database = new CohContentDatabase()
+      const database = new CohContentDatabase(
+        bundleDataFixture.omit('servers').create(),
+      )
       expect(database.servers).toHaveLength(0)
     })
 
     test(`should load values from bundle`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture
-        .create({ servers: ['Foo', 'Bar'] }))
-
+      const database = new CohContentDatabase(
+        bundleDataFixture
+          .create({ servers: ['Foo', 'Bar'] }),
+      )
       expect(database.servers).toStrictEqual(['Foo', 'Bar'])
     })
   })
 
   describe('archetypes', () => {
-    test(`should throw an error on duplicate key`, () => {
-      const database = new CohContentDatabase()
-      expect(() => database.load(bundleDataFixture.create({
-        archetypes: [
-          archetypeDataFixture.create({ key: 'foo' }),
-          archetypeDataFixture.create({ key: 'foo' }),
-        ],
-      }))).toThrow(`Duplicate key [foo]`)
-    })
-
     test(`should accept an undefined field`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture
+      const database = new CohContentDatabase(bundleDataFixture
         .omit('archetypes')
         .create())
       expect(database.archetypes).toHaveLength(0)
     })
 
     test(`should load data from bundle`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture
-        .create({ archetypes: [archetypeDataFixture.create({ key: 'foo' })] }))
+      const database = new CohContentDatabase(
+        bundleDataFixture
+          .create({ archetypes: [archetypeDataFixture.create({ key: 'foo' })] }),
+      )
       expect(database.getArchetype('foo')).not.toBeUndefined()
     })
   })
 
   describe('badges', () => {
-    test(`should throw an error on duplicate key`, () => {
-      const database = new CohContentDatabase()
-      expect(() => database.load(bundleDataFixture.create({
-        badges: [
-          badgeDataFixture.create({ key: 'foo' }),
-          badgeDataFixture.create({ key: 'foo' }),
-        ],
-      }))).toThrow('Duplicate key [foo]')
+    test(`should accept an undefined field`, () => {
+      const database = new CohContentDatabase(
+        bundleDataFixture.omit('badges').create(),
+      )
+      expect(database.badges).toHaveLength(0)
     })
 
-    test(`should accept an undefined field`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture
-        .omit('badges')
-        .create())
-      expect(database.badges).toHaveLength(0)
+    test(`should load data from bundle`, () => {
+      const database = new CohContentDatabase(
+        bundleDataFixture
+          .create({ badges: [badgeDataFixture.create({ key: 'foo' })] }),
+      )
+      expect(database.getBadge('foo')).not.toBeUndefined()
     })
   })
 
   describe('zones', () => {
-    test(`should throw an error on duplicate zone`, () => {
-      const database = new CohContentDatabase()
-      expect(() => database.load(bundleDataFixture.create({
-        zones: [
-          zoneDataFixture.create({ key: 'foo' }),
-          zoneDataFixture.create({ key: 'foo' }),
-        ],
-      }))).toThrow(`Duplicate key [foo]`)
+    test(`should accept an undefined field`, () => {
+      const database = new CohContentDatabase(
+        bundleDataFixture.omit('zones').create(),
+      )
+      expect(database.zones).toHaveLength(0)
     })
 
-    test(`should accept an undefined field`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture
-        .omit('zones')
-        .create())
-      expect(database.zones).toHaveLength(0)
+    test(`should load data from bundle`, () => {
+      const database = new CohContentDatabase(
+        bundleDataFixture
+          .create({ zones: [zoneDataFixture.create({ key: 'foo' })] }),
+      )
+      expect(database.getZone('foo')).not.toBeUndefined()
     })
   })
 
   describe('contacts', () => {
-    test(`should throw an error on duplicate contact`, () => {
-      const database = new CohContentDatabase()
-      expect(() => database.load(bundleDataFixture.create({
-        contacts: [
-          contactDataFixture.create({ key: 'foo' }),
-          contactDataFixture.create({ key: 'foo' }),
-        ],
-      }))).toThrow(`Duplicate key [foo]`)
+    test(`should accept an undefined field`, () => {
+      const database = new CohContentDatabase(
+        bundleDataFixture.omit('contacts').create(),
+      )
+      expect(database.contacts).toHaveLength(0)
     })
 
-    test(`should accept an undefined field`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture
-        .omit('contacts')
-        .create())
-      expect(database.contacts).toHaveLength(0)
+    test(`should load data from bundle`, () => {
+      const database = new CohContentDatabase(
+        bundleDataFixture
+          .create({ contacts: [contactDataFixture.create({ key: 'foo' })] }),
+      )
+      expect(database.getContact('foo')).not.toBeUndefined()
     })
   })
 
   describe('missions', () => {
-    test(`should throw an error on duplicate mission`, () => {
-      const database = new CohContentDatabase()
-      expect(() => database.load(bundleDataFixture.create({
-        missions: [
-          missionDataFixture.create({ key: 'foo' }),
-          missionDataFixture.create({ key: 'foo' }),
-        ],
-      }))).toThrow(`Duplicate key [foo]`)
+    test(`should accept an undefined field`, () => {
+      const database = new CohContentDatabase(
+        bundleDataFixture.omit('missions').create(),
+      )
+      expect(database.missions).toHaveLength(0)
     })
 
-    test(`should accept an undefined field`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture
-        .omit('missions')
-        .create())
-      expect(database.missions).toHaveLength(0)
+    test(`should load data from bundle`, () => {
+      const database = new CohContentDatabase(
+        bundleDataFixture
+          .create({ missions: [missionDataFixture.create({ key: 'foo' })] }),
+      )
+      expect(database.getMission('foo')).not.toBeUndefined()
     })
   })
 
   describe('getArchetype', () => {
     test(`should retrieve archetype from the index`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture.create({
-        archetypes: [archetypeDataFixture.create({ key: 'foo' })],
-      }))
+      const database = new CohContentDatabase(
+        bundleDataFixture.create({
+          archetypes: [archetypeDataFixture.create({ key: 'foo' })],
+        }),
+      )
       expect(database.getArchetype('foo')).not.toBeUndefined()
     })
 
     test(`should return undefined for unknown archetype`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture.create({ archetypes: [] }))
+      const database = new CohContentDatabase(
+        bundleDataFixture.create({ archetypes: [] }),
+      )
       expect(database.getArchetype('foo')).toBeUndefined()
     })
 
     test(`should return undefined for undefined key`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture.create({ archetypes: [] }))
+      const database = new CohContentDatabase(
+        bundleDataFixture.create({ archetypes: [] }),
+      )
       const key = undefined
       expect(database.getArchetype(key)).toBeUndefined()
     })
@@ -215,23 +163,23 @@ describe(CohContentDatabase.name, () => {
 
   describe('getZone', () => {
     test(`should retrieve zone from the index`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture.create({
-        zones: [zoneDataFixture.create({ key: 'foo' })],
-      }))
-
+      const database = new CohContentDatabase(
+        bundleDataFixture.create({
+          zones: [zoneDataFixture.create({ key: 'foo' })],
+        }),
+      )
       expect(database.getZone('foo')).not.toBeUndefined()
     })
 
     test(`should return undefined for unknown zone`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture.create({ zones: [] }))
+      const database = new CohContentDatabase(bundleDataFixture.create({ zones: [] }))
       expect(database.getZone('foo')).toBeUndefined()
     })
 
     test(`should return undefined for undefined key`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture.create({ zones: [] }))
+      const database = new CohContentDatabase(
+        bundleDataFixture.create({ zones: [] }),
+      )
       const key = undefined
       expect(database.getZone(key)).toBeUndefined()
     })
@@ -239,22 +187,25 @@ describe(CohContentDatabase.name, () => {
 
   describe('getContact', () => {
     test(`should retrieve contact from the index`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture.create({
-        contacts: [contactDataFixture.create({ key: 'foo' })],
-      }))
+      const database = new CohContentDatabase(
+        bundleDataFixture.create({
+          contacts: [contactDataFixture.create({ key: 'foo' })],
+        }),
+      )
       expect(database.getContact('foo')).not.toBeUndefined()
     })
 
     test(`should return undefined for unknown contact`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture.create({ contacts: [] }))
+      const database = new CohContentDatabase(
+        bundleDataFixture.create({ contacts: [] }),
+      )
       expect(database.getContact('foo')).toBeUndefined()
     })
 
     test(`should return undefined for undefined key`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture.create({ contacts: [] }))
+      const database = new CohContentDatabase(
+        bundleDataFixture.create({ contacts: [] }),
+      )
       const key = undefined
       expect(database.getContact(key)).toBeUndefined()
     })
@@ -262,22 +213,25 @@ describe(CohContentDatabase.name, () => {
 
   describe('getMission', () => {
     test(`should retrieve mission from the index`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture.create({
-        missions: [missionDataFixture.create({ key: 'foo' })],
-      }))
+      const database = new CohContentDatabase(
+        bundleDataFixture.create({
+          missions: [missionDataFixture.create({ key: 'foo' })],
+        }),
+      )
       expect(database.getMission('foo')).not.toBeUndefined()
     })
 
     test(`should return undefined for unknown mission`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture.create({ missions: [] }))
+      const database = new CohContentDatabase(
+        bundleDataFixture.create({ missions: [] }),
+      )
       expect(database.getMission('foo')).toBeUndefined()
     })
 
     test(`should return undefined for undefined key`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture.create({ missions: [] }))
+      const database = new CohContentDatabase(
+        bundleDataFixture.create({ missions: [] }),
+      )
       const key = undefined
       expect(database.getMission(key)).toBeUndefined()
     })
@@ -285,22 +239,25 @@ describe(CohContentDatabase.name, () => {
 
   describe('getBadge', () => {
     test(`should retrieve badge from the index`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture.create({
-        badges: [badgeDataFixture.create({ key: 'foo' })],
-      }))
+      const database = new CohContentDatabase(
+        bundleDataFixture.create({
+          badges: [badgeDataFixture.create({ key: 'foo' })],
+        }),
+      )
       expect(database.getBadge('foo')).not.toBeUndefined()
     })
 
     test(`should return undefined for unknown badge`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture.create({ badges: [] }))
+      const database = new CohContentDatabase(
+        bundleDataFixture.create({ badges: [] }),
+      )
       expect(database.getBadge('foo')).toBeUndefined()
     })
 
     test(`should return undefined for undefined key`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture.create({ badges: [] }))
+      const database = new CohContentDatabase(
+        bundleDataFixture.create({ badges: [] }),
+      )
       const key = undefined
       expect(database.getBadge(key)).toBeUndefined()
     })
@@ -308,13 +265,14 @@ describe(CohContentDatabase.name, () => {
 
   describe('searchBadges', () => {
     test(`should search the badge list`, () => {
-      const database = new CohContentDatabase()
-      database.load(bundleDataFixture.create({
-        badges: [
-          badgeDataFixture.create({ key: 'foo', name: [{ value: 'Foo' }] }),
-          badgeDataFixture.create({ key: 'bar', name: [{ value: 'Bar' }] }),
-        ],
-      }))
+      const database = new CohContentDatabase(
+        bundleDataFixture.create({
+          badges: [
+            badgeDataFixture.create({ key: 'foo', name: [{ value: 'Foo' }] }),
+            badgeDataFixture.create({ key: 'bar', name: [{ value: 'Bar' }] }),
+          ],
+        }),
+      )
 
       const result = database.searchBadges({ query: { str: 'oo' } })
       expect(result.totalItems).toBe(1)